% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{IMURUN_OPTIONAL}
\alias{IMURUN_OPTIONAL}
\title{Optional (recognized but not required) input columns}
\format{
An object of class \code{list} of length 3.
}
\usage{
IMURUN_OPTIONAL
}
\description{
Columns that 'imuGAP' understands but does not require. They are
included as headers in the shipped template so users know they are
available.
}
\keyword{internal}
