% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{IMURUN_SHEETS}
\alias{IMURUN_SHEETS}
\title{The three input/sheet names imurun expects}
\format{
An object of class \code{character} of length 3.
}
\usage{
IMURUN_SHEETS
}
\description{
The three input/sheet names imurun expects
}
\keyword{internal}
