% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\docType{data}
\name{IMURUN_TARGET_OPTIONAL}
\alias{IMURUN_TARGET_OPTIONAL}
\title{Optional (recognized but not required) target-request columns}
\format{
An object of class \code{character} of length 2.
}
\usage{
IMURUN_TARGET_OPTIONAL
}
\description{
Optional (recognized but not required) target-request columns
}
\keyword{internal}
